fixing code Design
