--- app/src/main/java/com/eventable/model/Question.kt
@@ -1,7 +1,0 @@
-package com.eventable.model
-
-data class Question (
-    var questionname : String = "",
-    var votesyes : Int = 0,
-    var votesno : Int = 0
-)
